+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
 
 ## Unreleased
+### Changed
+- handler-ec2_node.rb updated to allow configuration set from client config
 
 ## [2.1.1] - 2016-02-05
 ### Added